Travis CI
Refactorized directory structure, included .travis.yml
