+++ src/main/java/minesweeper/logic/BombException.java
